+++ src/commands/ApplicationCommand.ts
@@ -1,13 +1,14 @@
-import type { CommandInteraction } from 'discord.js'
+import type { ApplicationCommandOption, CommandInteraction } from 'discord.js'
 
 /** Metadata of an application command. */
 export interface CommandMetadata {
   name: string
   description: string
+  options: ApplicationCommandOption[]
 }
 
 /** A function to represent the logic of an application command. */
-export type CommandLogic = (interaction: CommandInteraction) => Promise<void>
+export type CommandLogic = (interaction: CommandInteraction) => Promise<any>
 
 /** A function to be called once all commands have been registered. */
 export type CommandOnLoad = (commands: CommandMetadata[]) => void