+++ docs/CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.0]: 2023-10-03
+
+### Added
+
+- Verse cooldowns: the bot will not post the same verse in the same version
+  within a default of 180 seconds (3 minutes) of the last time it was posted.
+- `/cooldown` command to show and change the verse cooldown. Setting it to 0
+  will disable the cooldown.
+
 ## [0.15.0]: 2023-09-24
 
 ### Added
@@ -279,6 +288,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bot goes online.
 
 [unreleased]: https://github.com/gimjb/pbdb/compare/latest...HEAD
+[0.16.0]: https://github.com/gimjb/pbdb/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/gimjb/pbdb/compare/v0.14.1...v0.15.0
 [0.14.1]: https://github.com/gimjb/pbdb/compare/v0.14.0...v0.14.1
 [0.14.0]: https://github.com/gimjb/pbdb/compare/v0.13.0...v0.14.0